Создал модели и контроллеры для LotteryGameMatch и LotteryGameMatchUser.Реализовал следующие API, учитывая ролевой доступ: Редактирование и удаление пользователя, получение списка всех лотерейных игр, создание и завершение лотерейного матча, запись игрока на лотерейную игру, получение списка всех матчей по id lottery game, получение списка всех пользователей.

+++ app/Http/Controllers/UserController.php
@@ -10,46 +10,65 @@ class UserController extends Controller
 {
     public function update(Request $request, $id)
     {
-        Auth::id();
-
-
-        if(User::where('id', $id)->exists()) {
-            $user = User::find($id);
-            $user->first_name = is_null($request->first_name) ? $user->first_name : $request->first_name;
-            $user->last_name = is_null($request->last_name) ? $user->last_name : $request->last_name;
-            $user->email = is_null($request->email) ? $user->email : $request->email;
-            $user->save();
-            return response()->json([
-                "message" => "User update."
-            ], 404);
+        if (Auth::id() == $id) {
+            if(User::where('id', $id)->exists()) {
+                $user = User::find($id);
+                $user->first_name = is_null($request->first_name) ? $user->first_name : $request->first_name;
+                $user->last_name = is_null($request->last_name) ? $user->last_name : $request->last_name;
+                $user->email = is_null($request->email) ? $user->email : $request->email;
+                $user->save();
+                return response()->json([
+                    "message" => "User update."
+                ], 202);
+            }
+            else {
+                return response()->json([
+                    "message" => "User not found."
+                ], 404);
+            }
         }
         else {
             return response()->json([
-                "message" => "User not found."
-            ], 404);
+                "message" => "User is not authorized."
+            ], 403);
         }
     }
 
     public function destroy($id)
     {
-        if(User::where('id', $id)->exists()) {
-            $user = User::find($id);
-            $user->delete();
+        if (Auth::id() == $id) {
+            if(User::where('id', $id)->exists()) {
+                $user = User::find($id);
+                $user->delete();
 
-            return response()->json([
-                "message" => "User deleted."
-            ], 202);
+                return response()->json([
+                    "message" => "User deleted."
+                ], 202);
+            }
+            else {
+                return response()->json([
+                    "message" => "User not found."
+                ], 404);
+            }
         }
         else {
             return response()->json([
-                "message" => "User not found."
-            ], 404);
+                "message" => "User is not authorized."
+            ], 403);
         }
     }
 
     public function getAll()
     {
-        $user = User::all();
-        return response()->json($user);
+        $admin = Auth::user();
+        if ($admin->is_admin) {
+            $user = User::all();
+            return response()->json($user);
+        }
+        else {
+            return response()->json([
+                "message" => "Not enough permissions."
+            ], 403);
+        }
     }
 }
